runfix: split handling quotes and replies
